Uncircumvent radio chip's status for IEEE 802.15.4 hardware
Currently the cc2420 and at86rf231 hardware addresses (IEEE 802.15.4
short and long) and frequency channel are read only from a global
variable in the driver.  This global variable is only set when the
user sets the address/channel though both parameters might be preset
by the hardware. This change lets the driver read this parameters
directly from the hardware.