don't use an arbitrary IP address for myaddr
Now this probably needs some more work as there can be multiple local IP addresses, but currently the address that is compared agains the destination of incomming packets is rather arbitrary.
Also an address is only considered foreign when it's first 112 bits match, but the last 8 bit don't match - if the incomming address is completely different from myaddr, it's still considered a local address as there is no other code path. I've left this as is to not break things, but I think this needs further discussion or better documentation.