core: msg: make messaging optional for threads (2)
That's my take on #703 from @kaspar030:
> This commit makes it possible to create threads with CREATE_NOMSG,
disabling their messaging abilities.

The difference in my implementation is, that I don't use a pointer to
the message table, but make the thread control block as big as needed.

Incidentally I fixed the usage of `dINT()` and `eINT()` in msg.c,
because it would have been more complicated not to.