core: give recommendations for thread priorities
I think that its not useful to have a variable number of priority levels
depending on the architecture. That makes porting applications on other
boards needlessly more complicated.

Also the priority levels are not overly obvious. In our examples we
only use `PRIORITY_MAIN`, `PRIORITY_MAIN ± 1` but not much else.

This PR adds the meaningful-named macros:
```c
 #define SCHED_PRIO_LEVELS 16
 #define PRIORITY_IDLE     15

 #define PRIORITY_LEAST    13
 #define PRIORITY_LOW      10
 #define PRIORITY_MAIN     7
 #define PRIORITY_HIGH     4
 #define PRIORITY_HIGHEST  1

 #define PRIORITY_UP       (-1)
 #define PRIORITY_DOWN     (+1)
```