Add basic travis-ci integration
This PR adds the basic integration for [Travis CI][tci].

The config file `.travis.yml` tells the server to setup the basic build
system, and run `compile_test.py`. If not all  examples and tests could be
built, then the Travis CI integration adds a warning to "merge" button on
the bottom of the Github page.

Of course this integration makes little sense until #1049 is resolved,
because new bugs and old bugs cannot be told apart. Also (of course)
only because everything still builds, a PR still can intruduce a
multitude of runtime errors.

This integrations is meant to work additionally to our Jenkings. Users
can activate the test in their own RIOT forks to run the test before
opening a PR to master.

  [tci]: https://travis-ci.org/