drivers: fix includes in rtc.h for MSP430
The comment and the actual guard in `rtc.h` did not match.
If (as in "implication") the MCU is an MSP, then `sys/time.h` must be
included, to have `time_t`, `struct timeval`, …

Including the header file in any case should be safe, so I dropped the
guard altogether.