arm_common: no needless `_gettimeofday()` warning
The `arm_common` syscalls implement `_gettimeofday()` for the benefit of
newlib. The syscall call only work if the `rtc` module or the `vtimer`
modules is transcluded. If neither module is used, a warning is printed
by means of `#warning`.

This warning is useless if the user does not invoke `gettimeofday()`.
Further this warning prevents the use of `-Werror`.

This PR puts the function in its own file, that will only be linked if
it was used. And if the function was used, then a link time error occur
if neither `rtc` nor `vtimer` was transcluded.