x86: slightly better debuggable bincode
While tracking down an error in #1253 I ameded newlib a bit to be
slightly better debuggable. Some lengthy macros were removed in favor of
proper functions.

The parameter passing was wrong before (because of a bug in newlib),
because the user provided CFLAGS were overwritten in the subfolders.
It is kinda strange that this did not cause errors.

The standalone version of the toolchain script was deleted, because it
is not reasonable to maintain two scripts.