destiny: when calling posix bind, the user will probably pass a struct sockaddr_in6 and its corresponding sizeof() return value. However, in further steps this struct is cast to a sockaddr6_t. This cast by itself is not dangerous, since the former struct encapsulates the sockaddr6_t plus an additional field. But in later steps, memcpy is used with the former sizeof() value, which was passed to the posix bind call.
Due to different sizes of struct sockaddr_in6 and sockaddr6_t,
the port value of the foreign address field gets overwritten.
But TCP expects the foreign address to be completely zero
in order to handle further connections with this socket.

destiny: simple copy and caching the struct.
A simple copy suffices here. In addition the socket is cached to serve multiple assignments.