wireshark: fix dissector plugin and expand documentation
The wireshark dissector plugin currently fails with a "Lua Error" when
trying to parse nativenet packets. This patch restores working behaviour
by commenting the offending line introduced in 1258675, which intention
was to account for padding in small packets. It further adds a comment
on how to edit the plugin for the case padding information is required.

The README.md now contains a usage note for configuration on fedora
which might also be helpful for users of other distributions.