sys/readline: Added curses/ncurses support to the native port, which is needed for the shell/readline implementation.
The usage of ncurses is needed to configure the terminal/shell, in which native applications are run.
This adds the usage of the LIBNCURSES and LIBTINFO to the native port.
The ncurses initialization and de-initialization is done in the native-uart0 startup handlers.
To enable the proper restart of native terminal-apps, the ENV variable is stored and re-used upon restart.