board/rf231usb: initial support for rf231usb
This adds support for the Atmel reference design RF231USB-RD. It was
tested against the dresden elektronik "Funk-Stick deRFusb" and should
work with all their 23E0x variants.

The design combines a ATSAM3S4B cpu with a AT86RF231 2.4Ghz radio
transceiver on a USB stick. It is optionally equipped with 2GB flash
memory.

All devices can be flashed with SAM-BA through USB or with openocd
through JTAG (if pin header present).