gnrc_rpl: fix dependencies
The dependency conditional for RPL was at the very end of
`Makefile.dep`, giving depending modules no chance to pull in their
dependencies.

Also it pulled in the deprecated module `net_help`, though it did not
use any of it's functions (it just included `inet_pton.h` for `AF_INET6`).