sc_ipv6_nc: Use more letters for type
I was myself a little confused when I saw `R` for a registered address,
since I interpreted it as "router" (though the node wasn't supposed to
be a router). After a wild goose chase for wrongly set flags through the
code I found that `R` stands for "REGISTERED". To prevent such a thing
in the future I opted for fixing this in the output.