sys/xtimer: Handle (almost) any frequency hardware timers
Set XTIMER_HZ in periph_conf.h to specify the xtimer hardware timer
frequency.

The xtimer system uses the hardware timer's own tick representation
internally but a thin wrapper layer is added as a shim between xtimer
core and the public xtimer API which is based around microseconds as
the time unit.

Initially supports only 1000000 Hz (and powers-of-two divisions of it)
and 32768 Hz, but can easily be expanded by adding a new case for
XTIMER_HZ inside xtimer.h