Merging mutex_lock and mutex_wait
mutex_lock function was meaningless. It checked the mutex and if it is locked then this function checked it again in mutex_wait. In this PR I merged two functions in one and changed the sequence of DEBUG messages.

mutex_wait was not used anywhere in the code, as far as I know.

mutex_unlock_and_sleep test OK.