rbuf: use bitmap for overlapping detection
The old code use linked lists of intervals to check overlapping 6LoWPAN
fragment. That results in failures of list cell allocation under heavy load.

This commit changes it to use bitmap. As fragment offset/size is in unit of
eight bytes, the bitmap requires only 20 bytes per rbuf, that is less than a
quarter of memory requirement for the old implementation.