conn: rename connection to connectivity
In the paper about RIOT [1] we used the wording “connectivity” for `conn` to
differentiate the terminology from the fact that `conn` supports both
connection-less and connection-oriented transport types.

This change applies the same wording to the documentation.

[1] O. Hahm, K. Schleiser, H. Petersen, M. Lenders, P. Kietzmann, E. Baccelli,
T.C. Schmidt, and M. Wählisch. RIOT: Reconsidering Operating Systems for Low-End
IoT Devices. May 2016.