cpu/samr21: rename samr21.h back to the original samr21g18a.h
Also add the "Top header" samr21.h from ASF 3.17. This requires a define
like __SAMR21G18A__. That define is set in boards/samr21_xpro

There is a correction of the doxygen exclude patterns for new samr21
cmsis includes.

This is step 2 in reorganizing cpu/samXNN to be able to add support for
SAMD21