gnrc_ipv6_nib: Initial API for an IPv6 specific FIB/NIB
There are three main problems with the current NDP implementation:

1.  The data structures being handled by the neighbor discovery are
    spread all over the codebase (the FIB, gnrc_ipv6_netif which also
    doubles as prefix list, gnrc_ipv6_nc which also doubles as default
    router list).
2.  The data structures are manipulated more or less unprotected,
    everyone can basically change anything in them.
3.  No central event-management for the 11+ event types required by NDP

This PR introduces a very encapsulated API for an IPv6-specific FIB
(called NIB to distinguish it from the already existing FIB). At least
(1) and (2) are solved with this! Due to the encapsulated design (3) is
also very easy to solve, e.g. with the [event timer promised](
https://github.com/kaspar030/RIOT/tree/evtimer) by @kaspar030 in #5269.