firmware_swapping: compile to an elf file
Allow compiling any firmware by using:

    make FIRMWARE=firmware_name

Main thing to make it work was adding 'romslot1' and 'romslot2' in the
bootloader linker script with the section names '.slot.1.*'
The bootloader is compiled with .o of the firmware slots.

To create the firmware slots .o files, I use objcopy to load the binaries at
sections called '.slot.1.*'.