MAGIC HERE SEE COMMENTS: Makefile.base: add support for non recursive makefiles
By defining 'NONRECURSIVE', multiple modules Makefile can be included in the
same makefile and handled seperately.

WARNING:
If for any reason, a module includes another Makefile before Makefile.base it
will break the including makefile magic. So maybe it would be good separate it
to a Makefile.base.head and Makefile.base.tail for this kind of modules.