tests/unittests: Fix error due to macOS sed incompatibility
Since most people™ seem to be developing on linux machines, they
commonly have gnu sed installed. In my experience with cross-platform
projects, it makes sense to require installation of gnu-sed on macOS
since dealing with the incompatibilites of sed is non-trivial.

This commit introduces such a check for the relic package.

For reference, the imcompatible sed call is located in
pkg/relic/fix-util_print_wo_args.sh (e.g. in 775e2071e)

This loosely relates to #6473.