tests: remove unnecessary gnrc_netdev_default include
The `driver_enc28j60` and `driver_encx24j600` applications exist to
test the network device drivers of this type. Including the *default*
driver of the board is thus non-sensical and might even be harmful for
the success of this test (if default device in initialized before the
device that is tested).

Additionally, does the removal of this module allow for some boards to
fit this into their memory (at least in my linker configuration), so I
removed them from the `BOARD_INSUFFICIENT_MEMORY` list.