at86rf2xx: Fix wrong state transition when leaving sleep state
SLEEP is 0xf, the bitwise OR of
(AT86RF2XX_STATE_RX_AACK_ON | AT86RF2XX_STATE_TX_ARET_ON) equals 0x1f,
so this state transition to PLL_ON was wrongly triggered when leaving
sleep for any mode which has 0x1X as its state code. The symptom is a
hang in the transceiver driver thread because the transceiver is not
listening to state transition commands while in sleep mode.