netdev: add send queue to support split-phase send operation
Before this commit, the gnrc_netdev->send function was called immediately when a SEND
message was received by netdev. This requires the radio driver to "block" the thread until
the radio is free to send a packet. This commit changes netdev so support split-phase
sending in the radio driver. The netdev thread now waits until the current send operation
finishes before issuing a new one. This requires it to queue messages to send until the
radio driver is no longer busy. This allows the event loop to handle other events, such
as servicing interrupts or receiving packets, while the radio is busy sending a packet.
Therefore, it has better concurrency characteristics than the previous model, which
requires the gnrc_netdev->send function to block the event loop.