at86rf2xx: implement software CSMA/link retries; fix concurrency bugs
Although software CSMA and link retries are provided by the AT86RF233 radio's
"Extended Operating Mode", the radio is unable to receive any frames in between
CSMA probes and link retries. This interacts very poorly with protocols that
require bidirectional communication. If two nodes try to send frames to each
other at the same time, the CSMA backoff happens as it should, but when one of
the nodes finally attempts to send a frame after a CSMA backoff, the other node
is still performing CSMA backoff and therefore is not listening. The packet is
therefore lost. Therefore, we implement CSMA and link retries in software, in
such a way that packets may be received between CSMA probes or link retries.
This functionality is enabled by defining the AT86Rf2XX_SOFTWARE_CSMA flag at
compile-time. As this operation is split-phase, it requires the send queue at the
netdev layer, introduced in the previous commit.

We also fix concurrency issues in the at86rf2xx driver that were exacerbated when
we introduced software CSMA and link retries. Concurrency bugs in at86rf2xx driver
may cause a sent packet to overwrite the frame buffer while a packet is being received,
causing outgoing packets to be "received" by the driver, and incoming packets to be
dropped AFTER the radio sends a link-layer acknowledgment. Our fix is to detect, in a
race-free way, if a frame is being received when a frame is sent, and to mark it as a
CSMA failure before overwriting the radio's frame buffer. Note that the fix only works
when software CSMA is enabled for the AT86RF2XX radio (the AT86RF2XX_SOFTWARE_CSMA flag).