cflags: Set -gz -gsplit-dwarf by default if supported by the compiler
There are recurring questions on the mailing lists and on Github
regarding the size of object files when compiled with debugging
information, which is enabled by default in the RIOT build system.
By setting these flags the debugging information will be compressed and
placed in a separate file with the extension .dwo, one for each object
.o file.

Using this feature requires some tool versions:
 - GCC 4.8+ (released 22 March 2013) or Clang more recent than 2013
 - binutils 2.23+ (released 26 March 2013)
 - GDB 7.6+ (released 26 April 2013)

The build system will test if the compiler supports the option and only
add it if it does not cause a syntax error in the command line.