sys/newlib_syscalls_riot_clock: Implement clock() using xtimer.
Newlib's clock() depends on times() (which has to be supplied by the os).
This patch add two new (sub)modules:
- newlib_syscalls_riot_clock uses xtimer to provide for the clock() function.
- newlib_stubs_clock provides a stub implementation for _times_r() in case xtimer
  is not available and a stub is acceptable.

The new modules `newlib_syscalls_riot` and `newlib_stubs` are meant to contain
the rest of the system calls that will be split from newlib_syscalls_default
by groups.

For ease of use, if xtimer is used, the newlib_syscalls_riot_clock is automatically
included. This can be prevented by explicitly adding newlib_stubs_clock to USEMODULE.

The test, tests/libc_time_h_relative, can be compiled with or without
USEMODULE=newlib_stubs_clock to see the difference.